requirements.txt
pyserial==3.5
pyusb==1.2.1
pynput==1.7.6
